+++ tests/st/bert/utils.py
@@ -17,13 +17,9 @@
 Functional Cells used in Bert finetune and evaluation.
 """
 import math
-import numpy as np
-import tinyms as ts
-from tinyms import Tensor
-from tinyms import primitives as P
 from tinyms.callbacks import Callback
 
-from mindspore.nn.learning_rate_schedule import LearningRateSchedule, PolynomialDecayLR, WarmUpLR
+
 
 class LossCallBack(Callback):
     """
@@ -54,79 +50,5 @@ class LossCallBack(Callback):
                                                                str(cb_params.net_outputs)))
 
 
-class BertLearningRate(LearningRateSchedule):
-    """
-    Warmup-decay learning rate for Bert network.
-    """
-    def __init__(self, learning_rate, end_learning_rate, warmup_steps, decay_steps, power):
-        super(BertLearningRate, self).__init__()
-        self.warmup_flag = False
-        if warmup_steps > 0:
-            self.warmup_flag = True
-            self.warmup_lr = WarmUpLR(learning_rate, warmup_steps)
-        self.decay_lr = PolynomialDecayLR(learning_rate, end_learning_rate, decay_steps, power)
-        self.warmup_steps = Tensor(np.array([warmup_steps]).astype(np.float32))
-
-        self.greater = P.Greater()
-        self.one = Tensor(np.array([1.0]).astype(np.float32))
-        self.cast = P.Cast()
-
-    def construct(self, global_step):
-        decay_lr = self.decay_lr(global_step)
-        if self.warmup_flag:
-            is_warmup = self.cast(self.greater(self.warmup_steps, global_step), ts.float32)
-            warmup_lr = self.warmup_lr(global_step)
-            lr = (self.one - is_warmup) * decay_lr + is_warmup * warmup_lr
-        else:
-            lr = decay_lr
-        return lr
 
 
-def _get_poly_lr(global_step, lr_init, lr_end, lr_max, warmup_steps, total_steps, poly_power):
-    """
-    generate learning rate array
-
-    Args:
-       global_step(int): current step
-       lr_init(float): init learning rate
-       lr_end(float): end learning rate
-       lr_max(float): max learning rate
-       warmup_steps(int): number of warmup epochs
-       total_steps(int): total epoch of training
-       poly_power(int): poly learning rate power
-
-    Returns:
-       np.array, learning rate array
-    """
-    lr_each_step = []
-    if warmup_steps != 0:
-        inc_each_step = (float(lr_max) - float(lr_init)) / float(warmup_steps)
-    else:
-        inc_each_step = 0
-    for i in range(total_steps):
-        if i < warmup_steps:
-            lr = float(lr_init) + inc_each_step * float(i)
-        else:
-            base = (1.0 - (float(i) - float(warmup_steps)) / (float(total_steps) - float(warmup_steps)))
-            lr = float(lr_max - lr_end) * (base ** poly_power)
-            lr = lr + lr_end
-            if lr < 0.0:
-                lr = 0.0
-        lr_each_step.append(lr)
-
-    learning_rate = np.array(lr_each_step).astype(np.float32)
-    current_step = global_step
-    learning_rate = learning_rate[current_step:]
-    return learning_rate
-
-
-
-def get_bert_thor_damping(damping_max=5e-2, damping_min=1e-6, damping_power=1.0, damping_total_steps=30000):
-    damping = _get_poly_lr(global_step=0, lr_init=0.0, lr_end=damping_min, lr_max=damping_max, warmup_steps=0,
-                           total_steps=damping_total_steps, poly_power=damping_power)
-    return Tensor(damping)
-
-def get_bert_thor_lr(lr_max=0.0034, lr_min=3.244e-05, lr_power=1.0, lr_total_steps=30000):
-    learning_rate = _get_poly_lr(global_step=0, lr_init=0.0, lr_end=lr_min, lr_max=lr_max, warmup_steps=0,
-                                 total_steps=lr_total_steps, poly_power=lr_power)
-    return Tensor(learning_rate)
